+++ raw/print_darwin.go
@@ -30,11 +30,11 @@ func (me *GetAttrIn) String() string { return "" }
 func (me *ReadIn) String() string {
 	return fmt.Sprintf("{Fh %d off %d sz %d %s L %d %s}",
 		me.Fh, me.Offset, me.Size,
-		FlagString(readFlagNames, int(me.ReadFlags), ""))
+		FlagString(readFlagNames, int64(me.ReadFlags), ""))
 }
 
 func (me *WriteIn) String() string {
 	return fmt.Sprintf("{Fh %d off %d sz %d %s}",
 		me.Fh, me.Offset, me.Size,
-		FlagString(writeFlagNames, int(me.WriteFlags), ""))
+		FlagString(writeFlagNames, int64(me.WriteFlags), ""))
 }
